fix: http requests 'get' & 'getById'
Я не заметил при сдаче, у меня, оказывается, не работали любые http запросы, связанные с get. В нынешней версии пофиксил. Надеюсь, бана за такое своевольное действие не будет :(

+++ kotiki-java/DAL/src/main/java/ru/itmo/kotiki/dao/CatDao.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.itmo.kotiki.model.Cat;
 
 @Repository
-public interface CatDao extends JpaRepository<Cat, Long> {
+public interface CatDao extends JpaRepository<Cat, Integer> {
 }
