+++ kotiki-java/DAL/src/main/java/ru/itmo/kotiki/dao/OwnerDao.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.itmo.kotiki.model.Owner;
 
 @Repository
-public interface OwnerDao extends JpaRepository<Owner, Long> {
+public interface OwnerDao extends JpaRepository<Owner, Integer> {
 }
